+++ config.php
@@ -40,6 +40,19 @@ $config = [
     'youtube' => 'http://www.youtube.com',
     'vimeo' => 'http://www.vimeo.com',
     'instagram' => 'http://www.instagram.com',
+
+    'name' => 'Boilerplate',
+    'address' => 'Lindenstraat 90B',
+    'zip' => '9100',
+    'city' => 'Sint-Niklaas',
+    'phone' => '+32 (0)472 76 74 53',
+    'email' => 'mayday@maneuver.be',
+
+    'location' => (object)[
+      'address' => 'Lindenstraat 90B 9100 Sint-Niklaas',
+      'lat' => 51.221351,
+      'lng' => 4.285173,
+    ]
   ],
 
 ];